+++ android/Sora/Sora/src/main/java/jp/shiguredo/sora/audiomanager/SoraAudioManagerLegacy.java
@@ -25,7 +25,7 @@ class SoraAudioManagerLegacy extends SoraAudioManagerBase {
     private static final String TAG = "SoraAudioManagerLegacy";
     private enum AudioDevice { SPEAKER_PHONE, WIRED_HEADSET, EARPIECE, BLUETOOTH, NONE }
     private final AudioDevice defaultAudioDevice;
-    private final SoraBluetoothManager bluetoothManager;
+    private final SoraAudioManagerBluetooth bluetoothManager;
     private Set<AudioDevice> audioDevices = new HashSet<>();
     private boolean running;
     private boolean savedIsSpeakerPhoneOn;
@@ -40,7 +40,7 @@ class SoraAudioManagerLegacy extends SoraAudioManagerBase {
     private SoraAudioManagerLegacy(Context context) {
         super(context);
 
-        bluetoothManager = SoraBluetoothManager.create(context, this, audioManager);
+        bluetoothManager = SoraAudioManagerBluetooth.create(context, this, audioManager);
 
         // デフォルトのデバイスを設定する
         // 受話用のスピーカーがある場合は受話用のスピーカーを使う
@@ -133,9 +133,9 @@ class SoraAudioManagerLegacy extends SoraAudioManagerBase {
             return;
         }
 
-        if (bluetoothManager.getState() == SoraBluetoothManager.State.HEADSET_AVAILABLE
-                || bluetoothManager.getState() == SoraBluetoothManager.State.HEADSET_UNAVAILABLE
-                || bluetoothManager.getState() == SoraBluetoothManager.State.SCO_DISCONNECTING) {
+        if (bluetoothManager.getState() == SoraAudioManagerBluetooth.State.HEADSET_AVAILABLE
+                || bluetoothManager.getState() == SoraAudioManagerBluetooth.State.HEADSET_UNAVAILABLE
+                || bluetoothManager.getState() == SoraAudioManagerBluetooth.State.SCO_DISCONNECTING) {
             // Bluetooth のデバイスを更新する
             bluetoothManager.updateDevice();
         }
@@ -143,9 +143,9 @@ class SoraAudioManagerLegacy extends SoraAudioManagerBase {
         // 存在するオーディオデバイスのリストを生成する
         Set<AudioDevice> newAudioDevices = new HashSet<>();
 
-        if (bluetoothManager.getState() == SoraBluetoothManager.State.SCO_CONNECTED
-                || bluetoothManager.getState() == SoraBluetoothManager.State.SCO_CONNECTING
-                || bluetoothManager.getState() == SoraBluetoothManager.State.HEADSET_AVAILABLE) {
+        if (bluetoothManager.getState() == SoraAudioManagerBluetooth.State.SCO_CONNECTED
+                || bluetoothManager.getState() == SoraAudioManagerBluetooth.State.SCO_CONNECTING
+                || bluetoothManager.getState() == SoraAudioManagerBluetooth.State.HEADSET_AVAILABLE) {
             // Bluetooth デバイスが存在する
             newAudioDevices.add(AudioDevice.BLUETOOTH);
         }
@@ -193,20 +193,23 @@ class SoraAudioManagerLegacy extends SoraAudioManagerBase {
 
         // Bluetooth audio を開始する必要があるか
         boolean needBluetoothAudioStart =
-                bluetoothManager.getState() == SoraBluetoothManager.State.HEADSET_AVAILABLE
+                bluetoothManager.getState() == SoraAudioManagerBluetooth.State.HEADSET_AVAILABLE
                         && !isSetHandsfree
                         && (lastConnectedAudioDevice == AudioDevice.BLUETOOTH || !hasWiredHeadset);
-        Log.d(TAG, "Update device state: "
-                + "wired headset=" + hasWiredHeadset + ", "
-                + "BT state=" + bluetoothManager.getState() + ", "
-                + "need BT audio start=" + needBluetoothAudioStart);
 
         // Bluetooth audio を停止する必要があるか
         boolean needBluetoothAudioStop =
-                (bluetoothManager.getState() == SoraBluetoothManager.State.SCO_CONNECTED
-                        || bluetoothManager.getState() == SoraBluetoothManager.State.SCO_CONNECTING)
+                (bluetoothManager.getState() == SoraAudioManagerBluetooth.State.SCO_CONNECTED
+                        || bluetoothManager.getState() == SoraAudioManagerBluetooth.State.SCO_CONNECTING)
                         && (isSetHandsfree
                         || (lastConnectedAudioDevice == AudioDevice.WIRED_HEADSET && hasWiredHeadset));
+
+        Log.d(TAG, "Update device state: "
+                + "set handsfree=" + isSetHandsfree + ", "
+                + "wired headset=" + hasWiredHeadset + ", "
+                + "BT state=" + bluetoothManager.getState() + ", "
+                + "need BT audio start=" + needBluetoothAudioStart + ", "
+                + "need BT audio stop=" + needBluetoothAudioStop);
 
         if (needBluetoothAudioStop) {
             bluetoothManager.stopScoAudio();
@@ -214,7 +217,7 @@ class SoraAudioManagerLegacy extends SoraAudioManagerBase {
         }
 
         if (needBluetoothAudioStart && !needBluetoothAudioStop) {
-            // Bluetooth SCO audio を介しする
+            // Bluetooth SCO audio を開始する
             if (!bluetoothManager.startScoAudio()) {
                 // 失敗した場合はリストから BLUETOOTH を削除する
                 audioDevices.remove(AudioDevice.BLUETOOTH);
@@ -224,10 +227,10 @@ class SoraAudioManagerLegacy extends SoraAudioManagerBase {
 
         // ハンズフリーの解除を待つのは SCO_CONNECTING の間だけ
         willOffHandsfree = willOffHandsfree
-                && bluetoothManager.getState() == SoraBluetoothManager.State.SCO_CONNECTING;
+                && bluetoothManager.getState() == SoraAudioManagerBluetooth.State.SCO_CONNECTING;
 
         final AudioDevice newAudioDevice;
-        if (bluetoothManager.getState() == SoraBluetoothManager.State.SCO_CONNECTED) {
+        if (bluetoothManager.getState() == SoraAudioManagerBluetooth.State.SCO_CONNECTED) {
             newAudioDevice = AudioDevice.BLUETOOTH;
             willOffHandsfree = false;
         } else if (!isSetHandsfree && hasWiredHeadset) {
@@ -237,11 +240,11 @@ class SoraAudioManagerLegacy extends SoraAudioManagerBase {
         } else {
             newAudioDevice = defaultAudioDevice;
         }
+        setSpeakerphoneOn(newAudioDevice == AudioDevice.SPEAKER_PHONE);
         if (newAudioDevice != selectedAudioDevice) {
             Log.d(TAG, "New device status: "
                     + "available=" + audioDevices + ", "
                     + "selected=" + newAudioDevice);
-            setSpeakerphoneOn(newAudioDevice == AudioDevice.SPEAKER_PHONE);
             selectedAudioDevice = newAudioDevice;
             if (onChangeRouteObserver != null) {
                 onChangeRouteObserver.OnChangeRoute();