+++ test/connect_disconnect.cpp
@@ -50,7 +50,7 @@ class SoraClient : public std::enable_shared_from_this<SoraClient>,
         audio_track_id,
         pc_factory()->CreateAudioSource(cricket::AudioOptions()).get());
     video_track_ =
-        pc_factory()->CreateVideoTrack(video_track_id, video_source_.get());
+        pc_factory()->CreateVideoTrack(video_source_, video_track_id);
 
     ioc_.reset(new boost::asio::io_context(1));